+++ container-engine-lib/lib/backend_interface/objects/api_container/api_container_filters.go
@@ -1,13 +1,13 @@
-package engine
+package api_container
 
 import "github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/container_status"
 
-type EngineFilters struct {
-	// Disjunctive set of engine IDs to find engines for
+type APIContainerFilters struct {
+	// Disjunctive set of enclave IDs for which to return API containers
 	// If nil or empty, will match all IDs
-	IDs map[string]bool
+	EnclaveIDs map[string]bool
 
-	// Disjunctive set of statuses that returned engines must conform to
+	// Disjunctive set of statuses that returned API containers must conform to
 	// If nil or empty, will match all IDs
 	Statuses map[container_status.ContainerStatus]bool
 }