+++ container-engine-lib/docs/changelog.md
@@ -1,4 +1,10 @@
 # TBD
+### Features
+* Created a generic `ContainerStatus` object in the `KurtosisBackend` API for representing the state of containers
+
+### Breaking Changes
+* The `Engine` object's status will now be a `ContainerStatus`, rather than `EngineStatus`
+    * Users should migrate to the new version
 
 # 0.9.1
 ### Fixes